infer_phylogeny

This function calls Rmix from package Rphylip to perform the inference of the phylogeny with Camin-Sokal.

infer_phylogeny(
  REvoBC_object,
  phylip_package_path,
  mutations_use = "smooth_del"
)

Arguments

REvoBC_object

(Required)

phylip_package_path

(Required). Prior to running this function, users should install PHYLIP and then provide the path to the folder containing the executable of mix. For example, if in Windows a user stores the folder of phylip in path D:/Programs/phylip, then the value of phylip_package_path should be set to D:/Programs/phylip/exe/.

mutations_use

(optional). Default = 'smooth_del' A string indicating what type of utations to use for phylogeny reconstruction. Can be one of c('smooth_del', 'smooth_del_ins', sub_smooth_del_ins, sub_del_ins). The first corresponds to using only smoothed deletions. The second refers to the case where smoothed deletions and insertios are considered. The third considers the smoothed deletions/insertions and the substitutions. Finally the fourth uses all non-smoothed mutations.

Value

REvoBC object with a new filed called phylogeny, that stores the inferred tree.

Examples

if (FALSE) {
data(revo_msa)
output_dir = system.file("extdata", "output", package = "REvoBC")
revo_msa$output_directory = output_dir
revo_phyl = infer_phylogeny(revo_msa, phylip_package_path = 'D:/Programmi/phylip-3.698/exe/')
}