initialize_REvoBC

This function initializes the REvoBC object, by computing the set of Amplicon Sequence Variants. It wraps all the steps performed by dada2, that are the following: (1) Filter and trimming reads. (2) Learning the error rates. (3) Dereplication. (4) Sample Inference. (5) Merging forward and reverse reads. (6) Construction of the Amplicon Sequence Variant table (ASV). (7) Removal of chimeras. (8) [Optional] Production of a summary which tracks changes in the number of reads at every step of the pipeline.

initialize_REvoBC(
  output_dir,
  input_dir = NULL,
  dada2_output_sequences = NULL,
  output_dir_dada2 = NULL,
  random_seed = NULL,
  output_figures = TRUE,
  multithread = TRUE,
  map_file_sample = NULL,
  dada2_pooled_analysis = FALSE,
  dada2_chimeras_minFoldParentOverAbundance = 8,
  verbose = TRUE,
  ...
)

Arguments

output_dir
(Required). Path to the directory where all output files will be stored. The following .csv files will be created:

quality_track_reads.csv: track of the number of sequences during the different stepsof dada2 analysis.

dada2_asv_prefilter.csv: sequences detected by dada2, with the counts detected in each sample.

input_dir
Path to the directory containing .fastq files for forward and reverse reads. This folder should contain the fastq files (2 for each sample) with the following name pattern: FILEPREFIX_SAMPLE_BARCODEVERSION_R1.fastq FILEPREFIX_SAMPLE_BARCODEVERSION_R1.fastq. SAMPLE refers to either an organ (in case multiple organs were sequenced) or timepoint (if longitudinal data are provided). Note that REvoBC does not support mixed sample types (i.e. samples must be either all from organs or all from timepoints). Required when dada2_output_sequences is null (i.e. when the user has not previously run dada2).

dada2_output_sequences
(Optional). In case users have already run dada2 up to the removal of bimeras, they should provide the path to the csv file containing the output (See description). The output should contain samples on rows and sequences on columns.

output_dir_dada2
(Optional). Output folder for dada2 filtered fastqs. If it doesn't exist it is created. Default is NULL and results in creating a sub-folder in output_folder named 'filtered_fastq/'.

random_seed
(Optional). Seed for reproducibility. Deafult is NULL.

output_figures
(Optional). Boolean indicating whether or not to save intermediate dada2 plots such as read quality profiles. Deafult is TRUE and corresponds to creating a sub-folder inside output_dir named "dada2_plots". This folder will contain the following plots:

Quality profiles of both forward and reverse reads (generated by dada2).

Error rates (generated by dada2).

Histogram of sequences length.

multithread
(Optional) default TRUE. Whether to enable multithreading. If TRUE the number of threads is determined automatically (dada2 functionality). If an integer is given, the number of threads is determined by its value.

map_file_sample
(Optional). In case fastq files names are not in the format
FILEPREFIX_SAMPLE_FILESUFFIX_RX.fastq), then users should provide a list that associates each filename (without the suffix _R1 and _R2) to the corresponding organ/day. (e.g., if we have forward and reverse files named file1_R1.fastq and file1_R2.fastq that correspond to organ PRL (code for Prostate Left), than the parameter should be set as: map_file_sample = c("file1" = "PRL")).

dada2_pooled_analysis
(Optional). Deafault = FALSE. Boolean that is passed to dada2 function dada, which performs sample inference. It can be set to TRUE in case of multiple samples coming from the same mouse (e.g. different organs or multiple time points) (See here for more information).

dada2_chimeras_minFoldParentOverAbundance
(Optional). Deafult = 8. parameter passed to the dada2 function isBimeraDenovo during the call to dada2 removeBimeraDenovo.

verbose
(Optional). Default TRUE. Boolean indicating whether or not to print the text output of dada2 functions.

...
(Optional) Any additional parameters passed to dada2 functions.

Value

An object of type REvoBC, which is a list that will contain the following fields:

fastq_directory: directory where the input fastq files are located.

output_directory: directory where all the output files are being stored.

map_file_sample: dataframe has as many rows as the input datasets, and for each input stores the sample (e.g. organ or day for longitudinal data) to which it is associated.

dada2_asv_prefilter: dataframe that stores all sequences detected by dada2. Note that these sequences still need to be filtered (see also asv_analysis).

dada2: list which contains the percentage of chimeras found by dada2 and a dataframe that tracks the number of sequences during all dada2 steps.

. This function also saves the following .csv files in the subfolder dada2_files created in the output dirctory provided by the user:

dada2_asv_prefilter.csv: sequences detected by dada2, with the counts detected in each sample.

quality_track_reads.csv: track of the number of sequences during all dada2 steps.

Details

For further information please refer to the original vignette. If users wish to manually run each step of dada2 computation they should execute all steps reported in the vignette, up to the Removal of chimeras and save the output of removeBimeraDenovo in a csv file and provide the path to that file as parameter dada2_output_sequences.

Examples

input_dir = system.file("extdata", "input", package = "REvoBC")
output_dir = system.file("extdata", "output", package = "REvoBC")
initialize_REvoBC(input_dir = input_dir, output_dir = output_dir)
#> ℹ Found zipped fastq files, extracting
#> ℹ Done extracting
#> Found 2 fastq files
#> ℹ The folliwng files, mapped to the corresponding sample, were found: 
#> MMUS1469_HMR_BC10v0_MG_120419  :  HMR 
#> Working directory:  /tmp/RtmpPkbDxu/filef1866eab5fb0/referencePath:  /home/runner/work/_temp/Library/REvoBC/extdata/input
#> ℹ Creating quality profiles figures
#> Warning: `guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
#> Warning: `guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
#> ℹ Filtering and trimming
#> Read in 59148 paired-sequences, output 54321 (91.8%) filtered paired-sequences.
#> ℹ Learning errors for forward reads
#> 10908088 total bases in 54321 reads from 1 samples will be used for learning the error rates.
#> ℹ Learning errors for backward reads
#> 11131107 total bases in 54321 reads from 1 samples will be used for learning the error rates.
#> Warning: Transformation introduced infinite values in continuous y-axis
#> Warning: Transformation introduced infinite values in continuous y-axis
#> Warning: Transformation introduced infinite values in continuous y-axis
#> Warning: Transformation introduced infinite values in continuous y-axis
#> ℹ Dereplicating fastqs
#> Dereplicating sequence entries in Fastq file: /home/runner/work/_temp/Library/REvoBC/extdata/output/filtered_fastq/MMUS1469_HMR_BC10v0_MG_120419_F_filt.fastq.gz
#> Encountered 12705 unique sequences from 54321 total sequences read.
#> Dereplicating sequence entries in Fastq file: /home/runner/work/_temp/Library/REvoBC/extdata/output/filtered_fastq/MMUS1469_HMR_BC10v0_MG_120419_R_filt.fastq.gz
#> Encountered 14618 unique sequences from 54321 total sequences read.
#> ℹ Inferring samples
#> Sample 1 - 54321 reads in 12705 unique sequences.
#> Sample 1 - 54321 reads in 14618 unique sequences.
#> 40387 paired-reads (in 170 unique pairings) successfully merged out of 51398 (in 693 pairings) input.
#> Warning: Removed 2 rows containing missing values (geom_bar).
#> ℹ Removing bimeras
#> Identified 59 bimeras out of 170 input sequences.
#> $fastq_directory
#> [1] "/home/runner/work/_temp/Library/REvoBC/extdata/input"
#> 
#> $output_directory
#> [1] "/home/runner/work/_temp/Library/REvoBC/extdata/output"
#> 
#> $dada2
#> $dada2$track
#>     input filtered denoisedF denoisedR merged nonchim
#> HMR 59148    54321     53760     51812  40387   38851
#> 
#> $dada2$bimera_percentage
#> [1] 96.1968
#> 
#> $dada2$original_sequences
#> [1] 170
#> 
#> 
#> $map_file_sample
#>                               sample
#> MMUS1469_HMR_BC10v0_MG_120419    HMR
#> 
#> $dada2_asv_prefilter
#> # A tibble: 111 × 3
#>      HMR seq                                                           seq_names
#>    <int> <chr>                                                         <chr>    
#>  1 15731 TCTACACGCGCGTTCAACCGAGGAAAACTACACACACGTTCAACCACGGTTTTTTACACA… SEQ001   
#>  2  2669 TCTACACGCGCGTTCAACCGAGGAAAACTACACACGCATTCAACCACGGTTTATTACACG… SEQ002   
#>  3  2308 GCTCCCTTCCCTTCCCTTCTCTCCGGCCCGCGGCGGCCGCGCTCCTCACCTGGCCAGGGG… SEQ003   
#>  4  2157 TCTACACGCGCGTTCAACCGAGGAAAACTACACACACGTTCAACCACGGTTTGTTACACA… SEQ004   
#>  5  2121 TCTACACGCGCGTTCAACCGAGGAAAACTACACACACATTCAACCGTGGTTTACTACACA… SEQ005   
#>  6  2098 TCTACACGCGCGTTCAACCGAGGAAAACTACACACACATTCAACCGTGGTTTATTACACG… SEQ006   
#>  7  1443 CTGGGCGGCAACAATCCCATTGCGTGTC                                  SEQ007   
#>  8   896 GCTCCCTTCCCTTCCCTTCTCTCCGGCCCGCGGCGGCCGCGCTCCTCACCTGGCCAGGGG… SEQ008   
#>  9   724 CTGGCACTCGTTGCTGGGTCTGGGGACGTC                                SEQ009   
#> 10   608 TCTACACGCGCGTTCAACTACACACACGTTCAACCACGGTTTTTTACACACGCATTCAAC… SEQ010   
#> # … with 101 more rows
#> 
#> attr(,"class")
#> [1] "REvoBC"