perform_msa

This function performs multi-sequence alignment and outputs statistics about mutations. Multiple Sequence Alignment is performed through the MUSCLE algorithm.

perform_msa(REvoBC_object, ...)

Arguments

REvoBC_object
REvoBC object on which we want to perform msa.

...
Optional parameters (options and flags) passed to MUSCLE. See the original guide for detailed information on all possible options.

Value

REvoBC object with a new field named alignment, which is a list with the following fileds:

msa_stringset: output of MSA peformed with MUSCLE.

mutations_df: tibble where each line corresponds to a position in a ASV, and the columns encode the following information:

asv_names: name of the ASV

day_organ: sample

position_bc260: position of the alteration in the original barcode. Note that insertions are assigned to the position that coincides with their beginning.

alt: type of alteration. wt = Wild Type (i.e. non-mutated position). sub = substitution. del = deletion. ins = insertion.

perc_in_sample: out of all sequences that map to a sample, the percentage of them that display the alteration.

. In addition, the following files are saved in the sub-folder "msa" created inthe output directory chosen by the user:

dnastringset.fa: fasta file where the sequences are stored.

dnastringset.fa: same as fasta, but in csv format.

dnastringset_muscle-muscle_msa.fasta: fasta with the stringset resulted from MUSCLE.

ASV_alterationType_frequency.csv: number of alterations for each type in each ASV

mutations_frequency.csv: per sample normalized frequency of each alteration type, computed for each position of the barcode in each sample.

mutations_df.csv: content of mutations_df variable explained above.

Examples

data(revo_analyzed)
output_dir = system.file("extdata", "output", package = "REvoBC")
revo_analyzed$output_directory = output_dir
revo_msa = perform_msa(revo_analyzed)
#> Warning: attempt to set 'col.names' ignored
#> 
#> MUSCLE v3.8.31 by Robert C. Edgar
#> 
#> http://www.drive5.com/muscle
#> This software is donated to the public domain.
#> Please cite: Edgar, R.C. Nucleic Acids Res 32(5), 1792-97.
#> 
#> filee3133eaafc5f 21 seqs, max length 260, avg  length 175
#> 1625 MB(23%)00:00:00                Iter   1    0.43%  K-mer dist pass 11625 MB(23%)00:00:00                Iter   1  100.00%  K-mer dist pass 1
#> 1625 MB(23%)00:00:00                Iter   1    0.43%  K-mer dist pass 21625 MB(23%)00:00:00                Iter   1  100.00%  K-mer dist pass 2
#> 1625 MB(23%)00:00:00                Iter   1    5.00%  Align node       1625 MB(23%)00:00:00                Iter   1   10.00%  Align node1625 MB(23%)00:00:00                Iter   1   15.00%  Align node1625 MB(23%)00:00:00                Iter   1   20.00%  Align node1625 MB(23%)00:00:00                Iter   1   25.00%  Align node1625 MB(23%)00:00:00                Iter   1   30.00%  Align node1625 MB(23%)00:00:00                Iter   1   35.00%  Align node1625 MB(23%)00:00:00                Iter   1   40.00%  Align node1625 MB(23%)00:00:00                Iter   1   45.00%  Align node1625 MB(23%)00:00:00                Iter   1   50.00%  Align node1625 MB(23%)00:00:00                Iter   1   55.00%  Align node1625 MB(23%)00:00:00                Iter   1   60.00%  Align node1625 MB(23%)00:00:00                Iter   1   65.00%  Align node1625 MB(23%)00:00:00                Iter   1   70.00%  Align node1625 MB(23%)00:00:00                Iter   1   75.00%  Align node1625 MB(23%)00:00:00                Iter   1   80.00%  Align node1625 MB(23%)00:00:00                Iter   1   85.00%  Align node1625 MB(23%)00:00:00                Iter   1   90.00%  Align node1625 MB(23%)00:00:00                Iter   1   95.00%  Align node1625 MB(23%)00:00:00                Iter   1  100.00%  Align node1625 MB(23%)00:00:00                Iter   1  100.00%  Align node
#> 1625 MB(23%)00:00:00                Iter   1    4.76%  Root alignment1625 MB(23%)00:00:00                Iter   1    9.52%  Root alignment1625 MB(23%)00:00:00                Iter   1   14.29%  Root alignment1625 MB(23%)00:00:00                Iter   1   19.05%  Root alignment1625 MB(23%)00:00:00                Iter   1   23.81%  Root alignment1625 MB(23%)00:00:00                Iter   1   28.57%  Root alignment1625 MB(23%)00:00:00                Iter   1   33.33%  Root alignment1625 MB(23%)00:00:00                Iter   1   38.10%  Root alignment1625 MB(23%)00:00:00                Iter   1   42.86%  Root alignment1625 MB(23%)00:00:00                Iter   1   47.62%  Root alignment1625 MB(23%)00:00:00                Iter   1   52.38%  Root alignment1625 MB(23%)00:00:00                Iter   1   57.14%  Root alignment1625 MB(23%)00:00:00                Iter   1   61.90%  Root alignment1625 MB(23%)00:00:00                Iter   1   66.67%  Root alignment1625 MB(23%)00:00:00                Iter   1   71.43%  Root alignment1625 MB(23%)00:00:00                Iter   1   76.19%  Root alignment1625 MB(23%)00:00:00                Iter   1   80.95%  Root alignment1625 MB(23%)00:00:00                Iter   1   85.71%  Root alignment1625 MB(23%)00:00:00                Iter   1   90.48%  Root alignment1625 MB(23%)00:00:00                Iter   1   95.24%  Root alignment1625 MB(23%)00:00:00                Iter   1  100.00%  Root alignment1625 MB(23%)00:00:00                Iter   1  100.00%  Root alignment
#> 1625 MB(23%)00:00:00                Iter   2    5.26%  Refine tree   1625 MB(23%)00:00:00                Iter   2   10.53%  Refine tree1625 MB(23%)00:00:00                Iter   2   15.79%  Refine tree1625 MB(23%)00:00:00                Iter   2   21.05%  Refine tree1625 MB(23%)00:00:00                Iter   2   26.32%  Refine tree1625 MB(23%)00:00:00                Iter   2   31.58%  Refine tree1625 MB(23%)00:00:00                Iter   2   36.84%  Refine tree1625 MB(23%)00:00:00                Iter   2   42.11%  Refine tree1625 MB(23%)00:00:00                Iter   2   47.37%  Refine tree1625 MB(23%)00:00:00                Iter   2   52.63%  Refine tree1625 MB(23%)00:00:00                Iter   2   57.89%  Refine tree1625 MB(23%)00:00:00                Iter   2   63.16%  Refine tree1625 MB(23%)00:00:00                Iter   2   68.42%  Refine tree1625 MB(23%)00:00:00                Iter   2   73.68%  Refine tree1625 MB(23%)00:00:00                Iter   2   78.95%  Refine tree1625 MB(23%)00:00:00                Iter   2   84.21%  Refine tree1625 MB(23%)00:00:00                Iter   2   89.47%  Refine tree1625 MB(23%)00:00:00                Iter   2   94.74%  Refine tree1625 MB(23%)00:00:00                Iter   2  100.00%  Refine tree1625 MB(23%)00:00:00                Iter   2  100.00%  Refine tree
#> 1625 MB(23%)00:00:00                Iter   2    4.76%  Root alignment1625 MB(23%)00:00:00                Iter   2    9.52%  Root alignment1625 MB(23%)00:00:00                Iter   2   14.29%  Root alignment1625 MB(23%)00:00:00                Iter   2   19.05%  Root alignment1625 MB(23%)00:00:00                Iter   2   23.81%  Root alignment1625 MB(23%)00:00:00                Iter   2   28.57%  Root alignment1625 MB(23%)00:00:00                Iter   2   33.33%  Root alignment1625 MB(23%)00:00:00                Iter   2   38.10%  Root alignment1625 MB(23%)00:00:00                Iter   2   42.86%  Root alignment1625 MB(23%)00:00:00                Iter   2   47.62%  Root alignment1625 MB(23%)00:00:00                Iter   2   52.38%  Root alignment1625 MB(23%)00:00:00                Iter   2   57.14%  Root alignment1625 MB(23%)00:00:00                Iter   2   61.90%  Root alignment1625 MB(23%)00:00:00                Iter   2   66.67%  Root alignment1625 MB(23%)00:00:00                Iter   2   71.43%  Root alignment1625 MB(23%)00:00:00                Iter   2   76.19%  Root alignment1625 MB(23%)00:00:00                Iter   2   80.95%  Root alignment1625 MB(23%)00:00:00                Iter   2   85.71%  Root alignment1625 MB(23%)00:00:00                Iter   2   90.48%  Root alignment1625 MB(23%)00:00:00                Iter   2   95.24%  Root alignment1625 MB(23%)00:00:00                Iter   2  100.00%  Root alignment1625 MB(23%)00:00:00                Iter   2  100.00%  Root alignment
#> 1625 MB(23%)00:00:00                Iter   2  100.00%  Root alignment
#> 1625 MB(23%)00:00:00                Iter   3    5.13%  Refine biparts1625 MB(23%)00:00:00                Iter   3    7.69%  Refine biparts1625 MB(23%)00:00:00                Iter   3   10.26%  Refine biparts1625 MB(23%)00:00:00                Iter   3   12.82%  Refine biparts1625 MB(23%)00:00:00                Iter   3   15.38%  Refine biparts1625 MB(23%)00:00:00                Iter   3   17.95%  Refine biparts1625 MB(23%)00:00:00                Iter   3   20.51%  Refine biparts1625 MB(23%)00:00:00                Iter   3   23.08%  Refine biparts1625 MB(23%)00:00:00                Iter   3   25.64%  Refine biparts1625 MB(23%)00:00:00                Iter   3   28.21%  Refine biparts1625 MB(23%)00:00:00                Iter   3   30.77%  Refine biparts1625 MB(23%)00:00:00                Iter   3   33.33%  Refine biparts1625 MB(23%)00:00:00                Iter   3   35.90%  Refine biparts1625 MB(23%)00:00:00                Iter   3   38.46%  Refine biparts1625 MB(23%)00:00:00                Iter   3   41.03%  Refine biparts1625 MB(23%)00:00:00                Iter   3   43.59%  Refine biparts1625 MB(23%)00:00:00                Iter   3   46.15%  Refine biparts1625 MB(23%)00:00:00                Iter   3   48.72%  Refine biparts1625 MB(23%)00:00:00                Iter   3   51.28%  Refine biparts1625 MB(23%)00:00:00                Iter   3   53.85%  Refine biparts1625 MB(23%)00:00:00                Iter   3   56.41%  Refine biparts1625 MB(23%)00:00:00                Iter   3   58.97%  Refine biparts1625 MB(23%)00:00:00                Iter   3   61.54%  Refine biparts1625 MB(23%)00:00:00                Iter   3   64.10%  Refine biparts1625 MB(23%)00:00:00                Iter   3   66.67%  Refine biparts1625 MB(23%)00:00:00                Iter   3   69.23%  Refine biparts1625 MB(23%)00:00:00                Iter   3   71.79%  Refine biparts1625 MB(23%)00:00:00                Iter   3   74.36%  Refine biparts1625 MB(23%)00:00:00                Iter   3   76.92%  Refine biparts1625 MB(23%)00:00:00                Iter   3   79.49%  Refine biparts1625 MB(23%)00:00:00                Iter   3   82.05%  Refine biparts1625 MB(23%)00:00:00                Iter   3   84.62%  Refine biparts1625 MB(23%)00:00:00                Iter   3   87.18%  Refine biparts1625 MB(23%)00:00:00                Iter   3   89.74%  Refine biparts1625 MB(23%)00:00:00                Iter   3   92.31%  Refine biparts1625 MB(23%)00:00:00                Iter   3   94.87%  Refine biparts1625 MB(23%)00:00:00                Iter   3   97.44%  Refine biparts1625 MB(23%)00:00:00                Iter   3  100.00%  Refine biparts1625 MB(23%)00:00:00                Iter   3  102.56%  Refine biparts1625 MB(23%)00:00:00                Iter   3  100.00%  Refine biparts
#> 1625 MB(23%)00:00:00                Iter   4    5.13%  Refine biparts1625 MB(23%)00:00:00                Iter   4    7.69%  Refine biparts1625 MB(23%)00:00:00                Iter   4   10.26%  Refine biparts1625 MB(23%)00:00:00                Iter   4   12.82%  Refine biparts1625 MB(23%)00:00:00                Iter   4   15.38%  Refine biparts1625 MB(23%)00:00:00                Iter   4   17.95%  Refine biparts1625 MB(23%)00:00:00                Iter   4   20.51%  Refine biparts1625 MB(23%)00:00:00                Iter   4   23.08%  Refine biparts1625 MB(23%)00:00:00                Iter   4   25.64%  Refine biparts1625 MB(23%)00:00:00                Iter   4   28.21%  Refine biparts1625 MB(23%)00:00:00                Iter   4   30.77%  Refine biparts1625 MB(23%)00:00:00                Iter   4   33.33%  Refine biparts1625 MB(23%)00:00:00                Iter   4   35.90%  Refine biparts1625 MB(23%)00:00:00                Iter   4   38.46%  Refine biparts1625 MB(23%)00:00:00                Iter   4   41.03%  Refine biparts1625 MB(23%)00:00:00                Iter   4   43.59%  Refine biparts1625 MB(23%)00:00:00                Iter   4   46.15%  Refine biparts1625 MB(23%)00:00:00                Iter   4   48.72%  Refine biparts1625 MB(23%)00:00:00                Iter   4   51.28%  Refine biparts1625 MB(23%)00:00:00                Iter   4   53.85%  Refine biparts1625 MB(23%)00:00:00                Iter   4   56.41%  Refine biparts1625 MB(23%)00:00:00                Iter   4   58.97%  Refine biparts1625 MB(23%)00:00:00                Iter   4   61.54%  Refine biparts1625 MB(23%)00:00:00                Iter   4   64.10%  Refine biparts1625 MB(23%)00:00:00                Iter   4   66.67%  Refine biparts1625 MB(23%)00:00:00                Iter   4   69.23%  Refine biparts1625 MB(23%)00:00:00                Iter   4   71.79%  Refine biparts1625 MB(23%)00:00:00                Iter   4   74.36%  Refine biparts1625 MB(23%)00:00:00                Iter   4   76.92%  Refine biparts1625 MB(23%)00:00:00                Iter   4   79.49%  Refine biparts1625 MB(23%)00:00:00                Iter   4   82.05%  Refine biparts1625 MB(23%)00:00:00                Iter   4   84.62%  Refine biparts1625 MB(23%)00:00:00                Iter   4   87.18%  Refine biparts1625 MB(23%)00:00:00                Iter   4   89.74%  Refine biparts1625 MB(23%)00:00:00                Iter   4   92.31%  Refine biparts1625 MB(23%)00:00:00                Iter   4   94.87%  Refine biparts1625 MB(23%)00:00:00                Iter   4   97.44%  Refine biparts1625 MB(23%)00:00:00                Iter   4  100.00%  Refine biparts1625 MB(23%)00:00:00                Iter   4  102.56%  Refine biparts1625 MB(23%)00:00:00                Iter   4  100.00%  Refine biparts
#> 1625 MB(23%)00:00:00                Iter   5    5.13%  Refine biparts1625 MB(23%)00:00:00                Iter   5    7.69%  Refine biparts1625 MB(23%)00:00:00                Iter   5   10.26%  Refine biparts1625 MB(23%)00:00:00                Iter   5   12.82%  Refine biparts1625 MB(23%)00:00:00                Iter   5   15.38%  Refine biparts1625 MB(23%)00:00:00                Iter   5   17.95%  Refine biparts1625 MB(23%)00:00:00                Iter   5  100.00%  Refine biparts
#> 1625 MB(23%)00:00:00                Iter   5  100.00%  Refine biparts
#> Warning: position_stack requires non-overlapping x intervals
#> Warning: position_stack requires non-overlapping x intervals