example data obtained from running function asv_analys, where the original barcode is identified and all remaining ASVs are aligned to it.

example data obtained from running function asv_analys, where the original barcode is identified and all remaining ASVs are aligned to it.

data(revo_analyzed)

Format

RevoBC object updated with the following fields: #'

clean_asv_dataframe

ASV sequences identified post-filtering (contamination removed, sequences with a similarity higher than pid_cutoff_nmbc to the original barcode aggregated to it and ASVs named in increasing order (ASV01, ASV02, etc.) according to their total counts.

barcode

Info about the barcode selected for the current analysis.

statistics

another list with the following sub-fileds:

asv_df_percentages

dataframe with six columns. asv_names is the name of the ASV. day_organ is the sample identifier (e.g. ID of an organ or, in case of longitudinal data, of the timepoint); count: total counts for a specific ASV in a specific sample; perc_in_sample: counts normalized to the total counts in the corresponding sample; perc_asv: conuts normalized to the total counts for the corresponding ASV; perc_fold_to_max: counts normalized to the maximum counts observed for the corresponding ASV in a sample.

asv_totalCounts

for each ASV, total counts and number of samples in which it was detected.

sample_totalCounts

for each sample, total counts and number of distinct ASVs detected.

asv_diversity_perSample

measures of clonal richness and measures of heterogeneity computed for each sample based on the ASVs detected.

asv_persample_frequency

counts for each ASV in each sample.

asv_persample_detection

binary matrix indicating whether a sequence has been detected in the corresponding sample.

asv_toBarcode_similarity

edit distance, percentage similarity and alignment score of each ASV compared to the original barcode.

Value

RevoBC object updated with statistics about the ASVs.