ASV_analysis

This function performs the analysis on ASV sequences identified by dada2. First, it removes possible contamination sequences (ASVs resulting from other barcodes). Then it performs pairwise alignment using Needleman-Wunsch global alignment algorithm implemented in function pairwiseAlignment in package Biostrings, aligning each sequence to the original barcode considered in the analysis (See the Biostrings documentation here for more details). Then it considers all those sequences exhibiting a similarity higher than pid_cutoff_nmbc with the original barcode as Non-Marking Guide Controls. Finally, it computes different statistics for the ASVS, storing: the relative frequency of all ASVs in each sample. the relative frequency of each ASV in the samples. the counts for each ASV normalized to the counts of the sample with maximum frequency the frequency of the different ASVs in each sample.

asv_analysis(
  REvoBC_object,
  barcode = "BC10v0.ORG",
  output_figures = TRUE,
  pid_cutoff_nmbc = 98,
  asv_count_cutoff = 2,
  ...
)

Arguments

REvoBC_object
(Required). Object of class REvoBC, result of the function initialize_REvoBC

barcode
(Required). String indicating the barcode used in the experiment.

output_figures
(Optional). Deafult TRUE: Boolean indicating whether a user whishes to store a figure indicating the number of ASV tracked during the different steps of the analysis.

pid_cutoff_nmbc
(Optional). Default to 98%. Percentage of similarity between a sequence and the original barcode. The ASVs with a similarity obve this threshold will be considered as original non-mutated sequences in the analysis.

asv_count_cutoff
(Optional). Default to 2. Minimum number of counts for an ASV to be considered in the statistics.

...
Any additional parameters passed to pairwiseAlignment from Biostrings (See description).

Value

The REvoBC object passed as a parameter with the following new fields:

clean_asv_dataframe: ASV sequences identified post-filtering (contamination removed, sequences with a similarity higher than pid_cutoff_nmbc to the original barcode aggregated to it and ASVs named in increasing order (ASV01, ASV02, etc.) according to their total counts.

barcode: Info about the barcode selected for the current analysis.

statistics: another list with the following sub-fileds:

asv_df_percentages: dataframe with six columns. asv_names is the name of the ASV. day_organ is the sample identifier (e.g. ID of an organ or, in case of longitudinal data, of the timepoint); count: total counts for a specific ASV in a specific sample; perc_in_sample: counts normalized to the total counts in the corresponding sample; perc_asv: conuts normalized to the total counts for the corresponding ASV; perc_fold_to_max: counts normalized to the maximum counts observed for the corresponding ASV in a sample.

asv_totalCounts: for each ASV, total counts and number of samples in which it was detected.

sample_totalCounts: for each sample, total counts and number of distinct ASVs detected.

asv_diversity_perSample: measures of clonal richness and measures of heterogeneity computed for each sample based on the ASVs detected.

asv_persample_frequency: counts for each ASV in each sample.

asv_persample_detection: binary matrix indicating whether a sequence has been detected in the corresponding sample.

asv_toBarcode_similarity: edit distance, percentage similarity and alignment score of each ASV compared to the original barcode.

It saves the following .csv files in a sub-folder asv_analysis of the main output folder:

sequences_barcode_mapping.csv: dataframe that stores, for each sequence, its counts in all samples. It also overwrites the sequence name (column seq_names) of those that exactly match any of the possible barcodes, using the barcode identifier. The sequences that don't match the barcode in the current analysis may be due to contamination.

clean_asv_dataframe.csv, asv_totalCounts.csv, sample_totalCounts.csv, asv_df_percentages.csv, asv_diversity_perSample.csv asv_persample_frequency.csv,

asv_persample_detection.csv, asv_diversity_perSample.csv and asv_toBarcode_similarity.csv: content of the corresponding variables with the same name described above.

Examples

data(revo_initialized)
output_dir = system.file("extdata", "output", package = "REvoBC")
revo_initialized$output_directory = output_dir
revo_analyzed = asv_analysis(REvoBC_object = revo_initialized, barcode = 'BC10v0.ORG')
#> Warning: Removed 1 rows containing missing values (geom_vline).